+++ arduino/Buzzer.cpp
@@ -89,9 +89,21 @@ const Buzzer::Melody CARD_FULL{CARD_FULL_SEQUENCE, 255};
 const uint8_t OK_SEQUENCE[] = {START, DAH, P, DAH, P, DAH, P_L,    DAH, P, DIT, P, DAH, P, 0};
 const Buzzer::Melody OK{OK_SEQUENCE, 255};
 
-// SVC = ... ...- -.-.
-const uint8_t SVC_SEQUENCE[] = {START, DIT, P, DIT, P, DIT, P_L,   DIT, P, DIT, P, DIT, P, DAH, P_L,   DAH, P, DIT, P, DAH, P, DIT, P, 0};
-const Buzzer::Melody SVC{SVC_SEQUENCE, 255};
+// BT = -... -
+const uint8_t BT_SEQUENCE[] = {START, DAH, P, DIT, P, DIT, P, DIT, P_L,   DAH, P,   0};
+const Buzzer::Melody BT{BT_SEQUENCE, 255};
+
+// WF = .-- ..-.
+const uint8_t WF_SEQUENCE[] = {START, DIT, P, DAH, P, DAH, P_L,     DIT, P, DIT, P, DAH, P, DIT, P, 0};
+const Buzzer::Melody WF{WF_SEQUENCE, 255};
+
+// .
+const uint8_t DIT_SEQUENCE[] = {START, DIT, P, 0};
+const Buzzer::Melody Dit{DIT_SEQUENCE, 255};
+
+// -
+const uint8_t DAH_SEQUENCE[] = {START, DAH, P, 0};
+const Buzzer::Melody Dah{DAH_SEQUENCE, 255};
 
 } //namespace Morse;
 
@@ -120,7 +132,22 @@ void Buzzer::SignalOk()
     _player.Play(Morse::OK);
 }
 
-void Buzzer::SignalService()
+void Buzzer::SignalBluetooth()
 {
-    _player.Play(Morse::SVC);
+    _player.Play(Morse::BT);
+}
+
+void Buzzer::SignalWifi()
+{
+    _player.Play(Morse::WF);
+}
+
+void Buzzer::SignalDit()
+{
+    _player.Play(Morse::Dit);
+}
+
+void Buzzer::SignalDah()
+{
+    _player.Play(Morse::Dah);
 }