+++ arduino/Context.cpp
@@ -63,6 +63,10 @@ int8_t Context::Setup()
     eeprom.read(ADDROF(_record_retain_days), &_record_retain_days, sizeof(_record_retain_days));
     if (_record_retain_days == 0xff)
         _record_retain_days = 1;
+#ifdef ESP32
+    eeprom.read(ADDROF(_wifi_ssid), reinterpret_cast<uint8_t *>(_wifi_ssid), sizeof(_wifi_ssid));
+    eeprom.read(ADDROF(_wifi_pass), reinterpret_cast<uint8_t *>(_wifi_pass), sizeof(_wifi_pass));
+#endif //ESP32
 
     // Restore current record
     _recorder.Setup(RecorderAccess::RECORD_ADJUSTED_START,
@@ -151,3 +155,27 @@ void Context::SetRecordRetainDays(uint8_t days)
     _record_retain_days = days;
     eeprom.write(ADDROF(_record_retain_days), &_record_retain_days, sizeof(_record_retain_days));
 }
+
+#ifdef ESP32
+
+void Context::SetWifiSsid(const char *ssid)
+{
+    auto length = strlen(ssid);
+    if (length > sizeof(_wifi_ssid))
+        length = sizeof(_wifi_ssid);
+    memcpy(_wifi_ssid, ssid, length);
+    _wifi_ssid[length - 1] = 0;
+    eeprom.write(ADDROF(_wifi_ssid), reinterpret_cast<uint8_t *>(_wifi_ssid), length);
+}
+
+void Context::SetWifiPass(const char *pass)
+{
+    auto length = strlen(pass);
+    if (length > sizeof(_wifi_pass))
+        length = sizeof(_wifi_pass);
+    memcpy(_wifi_pass, pass, length);
+    _wifi_pass[length - 1] = 0;
+    eeprom.write(ADDROF(_wifi_pass), reinterpret_cast<uint8_t *>(_wifi_pass), length);
+}
+
+#endif //ESP32