+++ arduino/CpuFreq.cpp
@@ -1,0 +1,14 @@
+#ifdef ESP32
+
+#include "CpuFreq.h"
+#include <Arduino.h>
+
+void SetCpuFreq(int mhz)
+{
+    setCpuFrequencyMhz(mhz);
+    Serial.print(F("CPU Frequency: "));
+    Serial.print(getCpuFrequencyMhz());
+    Serial.println(F(" MHz"));
+}
+
+#endif //ESP32