+++ arduino/CpuFreq.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#ifdef ESP32
+
+void SetCpuFreq(int mhz);
+
+#endif //ESP32