+++ arduino/src/Utils.cpp
@@ -1,0 +1,14 @@
+#include "Utils.h"
+
+namespace AOP {
+
+char* PrintNum(uint8_t num, char *buf)
+{
+    if (!num)
+        return buf;
+    buf = PrintNum(num / 10, buf);
+    *buf = '0' + num % 10;
+    return ++buf;
+}
+
+} //namespace AOP;